Add a sample code for Documaster's integration to eByggesak.
